federated-module/federated_model_training/tensorflow/requirements.txt
kafka-python
numpy<1.24
scikit-learn
seaborn
confluent-kafka
web3
protobuf
tensorflow==2.12.0
tensorflow-io==0.32.0
